Move log parsing functionality to edgelet core.
