+++ edgelet/edgelet-core/src/lib.rs
@@ -16,6 +16,7 @@ mod certificate_properties;
 pub mod crypto;
 mod error;
 mod identity;
+pub mod logs;
 mod module;
 mod network;
 mod settings;
@@ -32,6 +33,7 @@ pub use crypto::{
 };
 pub use error::{Error, ErrorKind};
 pub use identity::{AuthType, Identity, IdentityManager, IdentityOperation, IdentitySpec};
+pub use logs::{Chunked, LogDecode, LogChunk};
 pub use module::{
     ImagePullPolicy, LogOptions, LogTail, MakeModuleRuntime, Module, ModuleOperation,
     ModuleRegistry, ModuleRuntime, ModuleRuntimeErrorReason, ModuleRuntimeState, ModuleSpec,
